Reorder and restyling
