+++ lib/app.dart
@@ -5,7 +5,6 @@ import 'localizations.dart';
 import 'pages/calendar.dart';
 import 'pages/diagnose.dart';
 import 'pages/docs.dart';
-import 'pages/login.dart';
 import 'pages/maps.dart';
 import 'pages/news.dart';
 import 'pages/reporting.dart';
@@ -31,11 +30,9 @@ class MyApp extends StatelessWidget {
       ),
       initialRoute: '/',
       routes: {
-        '/': (context) => LoginPage(title: AppLoc.of(context).titleLoginPage),
+        '/': (context) => PortalPage(title: AppLoc.of(context).titlePortalPage),
         '/reporting': (context) =>
             ReportingPage(title: AppLoc.of(context).titleReportingPage),
-        '/login': (context) =>
-            LoginPage(title: AppLoc.of(context).titleLoginPage),
         '/diagnose': (context) =>
             DiagnosePage(title: AppLoc.of(context).titleDiagnosePage),
         '/docs': (context) =>
@@ -44,8 +41,8 @@ class MyApp extends StatelessWidget {
             CalendarPage(title: AppLoc.of(context).titleCalendarPage),
         '/news': (context) => NewsPage(title: AppLoc.of(context).titleNewsPage),
         '/maps': (context) => MapsPage(title: AppLoc.of(context).titleMapsPage),
-        '/portal': (context) =>
-            PortalPage(title: AppLoc.of(context).titlePortalPage),
+        // '/portal': (context) =>
+        //     PortalPage(title: AppLoc.of(context).titlePortalPage),
       },
     );
   }
