+++ lib/components/page_animation.dart
@@ -1,25 +1,22 @@
-// Not used
-
 import 'package:flutter/material.dart';
 
-class CrossFadeRoute extends PageRouteBuilder {
+class SlideRightRoute extends PageRouteBuilder {
   final Widget widget;
 
-  CrossFadeRoute({this.widget})
+  SlideRightRoute({this.widget})
       : super(
-          pageBuilder: (BuildContext context, Animation<double> animation1,
-              Animation<double> animation2) {
-            return widget;
-          },
-          transitionsBuilder: (BuildContext context,
-              Animation<double> animation1,
-              Animation<double> animation2,
-              Widget child) {
-            return FadeTransition(
-              opacity: animation1,
-              child: child,
-            );
-          },
-          transitionDuration: Duration(milliseconds: 2000),
-        );
+            pageBuilder: (BuildContext context, Animation<double> animation1,
+                    Animation<double> animation2) =>
+                widget,
+            transitionsBuilder: (BuildContext context,
+                    Animation<double> animation1,
+                    Animation<double> animation2,
+                    Widget child) =>
+                SlideTransition(
+                  position: Tween<Offset>(
+                    begin: const Offset(1.0, 0.0),
+                    end: Offset.zero,
+                  ).animate(animation1),
+                  child: child,
+                ));
 }
